Free temporary star verts after creating vertexbuffer
Big module to render text.  Just a simple one line string right now but want to eventually fit to a rect and have adjusted origins and such

Notes on heat

Load font and some test text.  Stuff isn't working yet and not sure if it is drawing backwards or getting clipped by the near plane.  Stopped for overwatch league.