Many many hours of painful debugging to get text rendering.  Turned out to be a goofy zw value in the texcoords sent to the pixel shader.
I probably now have alot of redundant experimental states and such scattered everywhere as I was attempting random stuff to get something to draw

Attempted to bump over the 2 drawn stuff a little bit.  The docs say this is needed but it looks worse with it.  I've asked the directx discord about it but this is from 20 years ago so who would remember?

I finished a todo!