The pitch/yaw/roll method was going haywire for ships.  I remembered this problem from ages ago when I worked on a ship game called Flyboyz.  I sort of had an idea of what to do but discovered that there's no way to rotate a vector by a quaternion in the old xbox math libs.
I ended up stealing the one out of CGLM and it works.  I accumulate rotations about side/up vectors I determine at the time of rotation.  Seems to work!

The only problem now is figuring out to get a heading for the HUD.  I think I can just dot with 0 0 1 and arc cosine that and convert to degrees?